Organize SetupBTScreen into multiple files
