+++ lib/bluetooth/select_service.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+import 'package:marklin_bluetooth/bluetooth/bluetooth.dart';
+import 'package:marklin_bluetooth/widgets.dart';
+
+class SelectServiceScreen extends StatelessWidget {
+  const SelectServiceScreen({Key? key, this.onServiceSelected})
+      : super(key: key);
+
+  final Function(String serviceID, BluetoothService service)? onServiceSelected;
+
+  @override
+  Widget build(BuildContext context) {
+    assert(Bluetooth.device != null); // Needs connected BT device
+
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("Select Bluetooth Service"),
+      ),
+      body: FutureBuilder<List<BluetoothService>>(
+        future: Bluetooth.device!.discoverServices(),
+        initialData: const [],
+        builder: niceAsyncBuilder(
+          loadingText: "Getting services...",
+          activeBuilder: (BuildContext c, snapshot) {
+            return ListView(
+              children: snapshot.data!
+                  .map((serv) => TextTile(
+                        title: serv.uuid.toString(),
+                        onTap: () {
+                          onServiceSelected?.call(serv.uuid.toString(), serv);
+                        },
+                      ))
+                  .toList(),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
