+++ lib/controller.dart
@@ -3,8 +3,8 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
-import 'package:marklin_bluetooth/bluetooth.dart';
-import 'package:marklin_bluetooth/setup_bt_screen.dart';
+import 'package:marklin_bluetooth/bluetooth/bluetooth.dart';
+import 'package:marklin_bluetooth/bluetooth/setup_bluetooth_screen.dart';
 import 'package:marklin_bluetooth/firebase/races.dart';
 import 'package:marklin_bluetooth/widgets.dart';
